wmi::query

Function quote_and_escape_wql_str

pub fn quote_and_escape_wql_str(s: impl AsRef<str>) -> String

Quote/escape a string for WQL.

2.2.1 WQL Query references DMTF-DSP0004 (“CIM”) which, in reading section “4.11.1 String Constants”, seems to only require that \ and " be escaped. It’s underspecified in CIM what happens with unicode values - perhaps \xNNNN escaping would be appropriate for a more general purpose CIM string escaping function - but in my testing on Windows 10.0.19041.572, it seems that such values do not need to be escaped for WQL, and are treated as their expected unicode values just fine.

assert_eq!(quote_and_escape_wql_str(r#"C:\Path\With"In Name"#), r#""C:\\Path\\With\"In Name""#);